Change CellDomain back to u64. Better for basic solver, so/so for queue solver.

+++ src/queue_solver.rs
@@ -2,7 +2,7 @@ use tracing::{instrument, trace};
 
 use crate::solver::{Puzzle, Tables};
 
-type CellDomain = u16;
+type CellDomain = u64;
 
 // ── LiveTuple ─────────────────────────────────────────────────────────────────
